Add Bugzilla API keys support
Bugzilla now supports API keys. If you set "api_key" on any request,
that API key is used. Usernames are optional when using API keys.
However, there is a valid_login API that can be used to determine if an
API key is valid and belongs to a specified user.

This commit adds support for API keys. We prefer API keys when multiple
credential types are specified. We also support calling valid_login if a
username is specified alongside an API key.

Signed-off-by: AutomatedTester <dburns@mozilla.com>