Use correct response HTTP code for Bugzilla error handling tests
Bugzilla does not return HTTP 200 if there was an error, so the tests
have been updated to more accurately reflect the real response. This is
important since if additional handling is added in the future to handle
the HTTP 500 case (where the response isn't valid JSON), we'll need to
ensure the existing BugsyException cases don't regress, since they are
also non-200 HTTP response codes.