Add the ability to set the cc list. Fixes #33
Property to add or remove people from the cc list.

To add people to the cc list
>>> bug.cc = "automatedtester@mozilla.com"

or
>>> bug.cc = ["automatedtester@mozilla.com", "dburns@mozilla.com"]

If you want to remove an email from the list, the last
character of
the email address nees to be a `-`. For Example:

>>> bug.cc = "automatedtester@mozilla.com-"

Removes an email.

You can mix adding and removing
>>> bug.cc = ["automatedtester@mozilla.com", "dburns@mozilla.com-"]