Adds the ability to set or remove dependent/blocking bugs Fixes #32
Property to add or remove dependent bugs.

Properties are bug.depends_on and bug.blocks. Example of depends below.

To add dependent bugs
>>> bug.depends_on = 145678

or
>>> bug.depends_on = [145678, 999999]

If you want to remove a depends on from the list, the last
character of the keyword  needs to be a `-`. For Example:

>>> bug.depends_on = "123456-"

You can mix adding and removing
>>> bug.depends_on =  ["99999", "123456-"]